pyproject.toml
[project]
name = "aiia-poc"
version = "0.1.0"
description = "AI Impact Assessment Tool – API backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "ollama>=0.4.0",
    "python-dotenv>=1.0.0",
]
